+++ editor/render/TextureImage.go
@@ -26,7 +26,7 @@ func TextureImage(label string, cache *graphics.TextureCache, key resource.Key, 
 			scaleFactor := float32(math.Min(float64(size.X/width), float64(size.Y/height)))
 			imageSize := imgui.Vec2{X: width * scaleFactor, Y: height * scaleFactor}
 
-			bufferSize := imgui.Vec2{X: float32((size.X - imageSize.X) / 2), Y: float32((size.Y - imageSize.Y) / 2)}
+			bufferSize := imgui.Vec2{X: (size.X - imageSize.X) / 2, Y: (size.Y - imageSize.Y) / 2}
 			imgui.SetCursorPos(bufferSize)
 
 			imgui.ImageV(textureID, imageSize, imgui.Vec2{}, uv,
